chore: add newrelic logging

+++ packages/backend/src/webPush.ts
@@ -1,11 +1,25 @@
 /// <reference lib="webworker" />
 
+import * as newrelic from "newrelic";
 import * as webpush from "web-push";
 import { MessageDescriptor } from "./i18n";
 import { logger } from "./logger";
 import { prisma } from "./prisma";
 
-export async function sendNotification(
+export const sendNotification = (
+  subscription: {
+    endpoint: string;
+    keys: { p256dh: string; auth: string };
+  },
+  subject: string,
+  payload?: { title: MessageDescriptor; data: NotificationOptions }
+) =>
+  newrelic.startBackgroundTransaction(
+    "sendNotification",
+    sendNotification_(subscription, subject, payload)
+  );
+
+async function sendNotification_(
   subscription: {
     endpoint: string;
     keys: { p256dh: string; auth: string };
@@ -14,7 +28,7 @@ export async function sendNotification(
   payload?: { title: MessageDescriptor; data: NotificationOptions }
 ) {
   try {
-    await webpush.sendNotification(
+    const result = await webpush.sendNotification(
       {
         endpoint: subscription.endpoint,
         keys: subscription.keys,
@@ -29,17 +43,18 @@ export async function sendNotification(
         },
       }
     );
+    logger.info(result, "Sent push notification");
   } catch (error) {
-    if (error instanceof webpush.WebPushError) {
-      if (error.statusCode === 410) {
-        await prisma.pushNotification.delete({
-          where: {
-            endpoint: subscription.endpoint,
-          },
-        });
+    if (error instanceof webpush.WebPushError && error.statusCode === 410) {
+      await prisma.pushNotification.delete({
+        where: {
+          endpoint: subscription.endpoint,
+        },
+      });
 
-        logger.info(subscription, "Deleted stale subscription");
-      }
+      logger.info(subscription, "Deleted stale subscription");
+    } else {
+      logger.error(error);
     }
   }
 }
